// src/node_service/data/zod_generated.ts
export const UserSchema = z.object({
  id: z.number().int(),
  username: z.string(),
  name: z.string(),
  isAdmin: z.boolean(),
  grannyId: z.number().int().nullable(),
})